modified:   src/App.js 	modified:   src/components/WorkoutDetails.js 	modified:   src/components/WorkoutForm.js 	modified:   src/index.css 	modified:   src/pages/Home.js

+++ src/components/WorkoutDetails.js
@@ -1,14 +1,26 @@
-import { useWorkoutsContext } from "../hooks/useWorkoutsContext"
-
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
+// imports
+import { useWorkoutsContext } from "../hooks/useWorkoutsContext"
+import { useAuthContext } from "../hooks/useAuthContext"
+
 const WorkoutDetails = ({workout}) => {
     const { dispatch } = useWorkoutsContext()
+    const { user } = useAuthContext()
     
     const handleClick = async () => {
+
+        // check for user
+        if (!user) {
+            return
+        }
+
         const response = await fetch('/api/workouts/' + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}` // note the back ticks ``
+            }
         })
 
         const json = await response.json()
